test: share the frame collector and waitFor across the task suites
`collectFrames` had three copies — one exported from apps/tasks, one
file-local in the data layer's task tests, and one the spawner tests
added. It moves down to `@virtool/data/test/frames`, which both packages
reach. `waitFor` had two copies inside apps/tasks and moves to
`src/testing/waitFor.ts`.
